fix: 폴더 구조 정리
Index페이지 수정 바로 게임화면이 보이게 함
ts 파일 정리
