+++ src/TS/ControlTower.ts
@@ -1,0 +1,5 @@
+import '../style.css';
+import '../TS/components/scale_game';
+import '../TS/components/move_btn';
+import '../TS/components/icon_btn';
+import '../TS/components/timer';
