+++ src/TS/components/createTimer.ts
