+++ src/TS/pages/room/south_room.ts
@@ -1,12 +1,12 @@
 // SouthRoom 클래스 - 남쪽 방 구성
 // 퍼즐 미니게임과 금고 키패드 미니게임 구현
 
-import { IRoom, IInventoryItem } from '../../types/type.ts';
-import { PuzzleGame } from '../minigame/south_minigame.ts';
-import { PasswordKeypad } from '../minigame/password_keypad.ts';
-import { CreateSearchBtn } from '../../utils/createSearchBtn.ts';
-import { showCluePopup } from '../../utils/showCluePopup.ts';
-import itemManagerInstance from '../../utils/itemManagerInstance.ts';
+import { IRoom, IInventoryItem } from '../../../types/type.ts';
+import { PuzzleGame } from '../../minigame/south_minigame.ts';
+import { PasswordKeypad } from '../../minigame/password_keypad.ts';
+import { CreateSearchBtn } from '../../components/createSearchBtn.ts';
+import { showCluePopup } from '../../../utils/showCluePopup.ts';
+import itemManagerInstance from '../../../utils/itemManagerInstance.ts';
 
 export class SouthRoom implements IRoom {
   private puzzleGame: PuzzleGame;
