+++ src/components/common/icon_btn.html
@@ -4,8 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="/src/style.css" rel="stylesheet" />
-    <script type="module" src="/src/TS/icon_btn.ts"></script>
-    <script type="module" src="/src/TS/minigame/east_minigames.ts"></script>
+    <script type="module" src="/src/TS/components/icon_btn.ts"></script>
     <title>Document</title>
   </head>
   <body>
