+++ src/components/common/move.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="/src/style.css" rel="stylesheet" />
-    <script type="module" src="/src/TS/move_btn.ts"></script>
+    <script type="module" src="/src/TS/components/move_btn.ts"></script>
     <title>search</title>
   </head>
   <body class="bg-black">
