--- src/components/common/search.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="ko-KR">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link href="/src/style.css" rel="stylesheet" />
-    <title>search</title>
-  </head>
-  <body class="relative min-h-screen bg-pink-500">
-    <h1>단서찾기 버튼 컴포넌트</h1>
-    <div id="search-btn-box" class="relative w-full min-h-screen"></div>
-  </body>
-</html>
