+++ src/components/common/timer.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
     <link href="/src/style.css" rel="stylesheet" />
-    <script type="module" src="/src/TS/timer.ts"></script>
+    <script type="module" src="/src/TS/components/timer.ts"></script>
   </head>
   <body class="bg-black">
     <div id="timer-box"></div>
