+++ src/pages/normal_end.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="icon" type="image/png" href="/favicon.png" />
     <title>노멀 엔딩</title>
-    <script type="module" src="../TS/normal_end.ts"></script>
+    <script type="module" src="../TS/pages/normal_end.ts"></script>
     <link rel="stylesheet" href="../style.css" />
   </head>
   <body class="ending-body bg-black flex items-center justify-center min-h-screen flex-col font-wanted-sans">
