+++ src/pages/play_ground.html
@@ -3,12 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link href="/src/style.css" rel="stylesheet" />
     <link rel="icon" type="image/png" href="/favicon.png" />
-    <script type="module" src="/src/TS/move_btn.ts"></script>
-    <script type="module" src="/src/TS/icon_btn.ts"></script>
-    <script type="module" src="/src/TS/timer.ts"></script>
-    <script type="module" src="/src/TS/scale_game.ts"></script>
+    <script type="module" src="/src/TS/ControlTower.ts"></script>
+
     <title>코드의 그림자 - 플레이 페이지</title>
     <style>
       html,
