+++ src/pages/start_page.html
@@ -6,7 +6,7 @@
     <link rel="icon" type="image/png" href="/favicon.png" />
     <title>코드의 그림자</title>
     <link rel="stylesheet" href="../style.css" />
-    <script type="module" src="../TS/start_page.ts"></script>
+    <script type="module" src="../TS/pages/start_page.ts"></script>
   </head>
   <body class="start-height min-h-screen bg-black">
     <!-- 페이지 전환용 오버레이 -->
