+++ src/pages/true_end.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="icon" type="image/png" href="/favicon.png" />
     <title>트루 엔딩</title>
-    <script type="module" src="../TS/true_end.ts"></script>
+    <script type="module" src="../TS/pages/true_end.ts"></script>
     <link rel="stylesheet" href="../style.css" />
   </head>
   <body class="ending-body bg-black flex items-center justify-center min-h-screen flex-col font-wanted-sans">
