+++ src/utils/itemManagerInstance.ts
@@ -1,4 +1,4 @@
-import ItemManager from './itemManager';
+import ItemManager from '../TS/components/itemManager';
 
 const itemManagerInstance = new ItemManager();
