+++ src/utils/room_manager.ts
@@ -1,8 +1,8 @@
 import { TRoomDirection, IRoom } from '../types/type.ts';
-import { NorthRoom } from './room/north_room.ts';
-import { SouthRoom } from './room/south_room.ts';
-import { EastRoom } from './room/east_room.ts';
-import { WestRoom } from './room/west_room.ts';
+import { NorthRoom } from '../TS/pages/room/north_room.ts';
+import { SouthRoom } from '../TS/pages/room/south_room.ts';
+import { EastRoom } from '../TS/pages/room/east_room.ts';
+import { WestRoom } from '../TS/pages/room/west_room.ts';
 
 export class RoomManager {
   private static rooms: Record<TRoomDirection, IRoom> = {
